Centralize Research member identity formatting
Move ResearchDiff body subject and API-to-body target alias formatting into ResearchMemberIdentity so DiffCommand and ResearchDiff share one canonical spelling path.
